refactor: 시큐리티 설정을 global.auth 패키지로 이동
