+++ src/main/java/com/bookspot/global/auth/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.bookspot.global;
+package com.bookspot.global.auth;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
