+++ src/test/java/com/bookspot/book/presentation/BookControllerTest_DetailSearch.java
@@ -1,7 +1,7 @@
 package com.bookspot.book.presentation;
 
 import com.bookspot.book.application.BookService;
-import com.bookspot.global.SecurityConfig;
+import com.bookspot.global.auth.SecurityConfig;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +11,6 @@ import org.springframework.context.annotation.Import;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
