+++ src/test/java/com/bookspot/library/presentation/LibraryControllerTest.java
@@ -1,6 +1,6 @@
 package com.bookspot.library.presentation;
 
-import com.bookspot.global.SecurityConfig;
+import com.bookspot.global.auth.SecurityConfig;
 import com.bookspot.library.application.LibraryService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +10,6 @@ import org.springframework.context.annotation.Import;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
